+++ lib/comprehensive_utils.dart
@@ -1,9 +1,10 @@
+/// A collection of classes and functions containing simple and complex performance-oriented tools
 library comprehensive_utils;
 
 export 'src/extensions/stream_extensions.dart';
 export 'src/streams/distinct_connectable_stream.dart';
 export 'src/streams/distinct_subject.dart';
-export 'src/streams/get_stream.dart';
-export 'src/streams/mini_stream.dart';
+//export 'src/streams/get_stream.dart';
+//export 'src/streams/mini_stream.dart';
 export 'src/tools/observable_timer.dart';
 export 'src/widgets/fluent_list_view.dart';