+++ lib/src/streams/distinct_connectable_stream.dart
@@ -16,14 +16,21 @@ abstract class _AbstractDistinctConnectableStream<T, S extends Subject<T>,
 class DistinctValueConnectableStream<T>
     extends _AbstractDistinctConnectableStream<T, DistinctSubject<T>,
         DistinctValueStream<T>> implements DistinctValueStream<T> {
-  DistinctValueConnectableStream(Stream<T> source,
-      {bool sync = false, bool Function(T, T)? equals})
-      : super(source, DistinctSubject<T>(sync: sync, equals: equals));
+  DistinctValueConnectableStream(
+    Stream<T> source, {
+    bool sync = false,
+    bool Function(T, T)? equals,
+  }) : super(source, DistinctSubject<T>(sync: sync, equals: equals));
 
-  DistinctValueConnectableStream.seeded(Stream<T> source, T seedValue,
-      {bool sync = false, bool Function(T, T)? equals})
-      : super(source,
-            DistinctSubject<T>.seeded(seedValue, sync: sync, equals: equals));
+  DistinctValueConnectableStream.seeded(
+    Stream<T> source,
+    T seedValue, {
+    bool sync = false,
+    bool Function(T, T)? equals,
+  }) : super(
+          source,
+          DistinctSubject<T>.seeded(seedValue, sync: sync, equals: equals),
+        );
 
   @override
   bool get hasValue => _subject.hasValue;